Run iterative predictive analysis, looping over each individual's data

Run iterative predictive analysis, looping over each individual's data

run_eval(
  model,
  data,
  ids = NULL,
  parameters = NULL,
  fixed = NULL,
  omega = NULL,
  iov = NULL,
  ruv = NULL,
  dictionary = list(),
  group = NULL,
  weights = NULL,
  weight_prior = 1,
  censor_covariates = TRUE,
  incremental = FALSE,
  threads = 1,
  progress = TRUE
)

Arguments

model

either a PKPDsim model object, or a string pointing to a PKPDsim-generated model library, e.g. pkvancothomson

data

NONMEM-style data.frame, or path to CSV file with NONMEM data

ids

optional, vector of subject IDs to run analysis on (by default runs analysis on all subjects in dataset)

parameters

list of parameters

fixed

fix a specific parameters, supplied as vector of strings

omega

between subject variability, supplied as vector specifiying the lower triangle of the covariance matrix of random effects

iov

a list specifying the required metadata for implementation of IOV, specifically the coefficient of variation (CV %) of the IOV for each parameter and a vector of bin separators. For example, list(cv = list(CL = 0.1, V = 0.2), bins = c(0, 24, 48, 9999))

ruv

residual error variability magnitude, specified as list.

dictionary

data dictionary. Optional, a named character vector that specifies the column names to be used from the dataset.

group

name of column in data that groups observations together in iterative flow. By default each observation will be its own "group", but this can be used to group peaks and troughs together, or to group observations on the same day together. Grouping will be done prior to running the analysis, so cannot be changed afterwards.

weights

vector of weights for error. Length of vector should be same as length of observation vector. If NULL (default), all weights are equal. Used in both MAP and NP methods. Note that `weights` argument will also affect residuals (residuals will be scaled too).

weight_prior

weighting of priors in relationship to observed data, default = 1

censor_covariates

with the proseval tool in PsN, there is “data leakage” (of future covariates data): since the NONMEM dataset in each step contains the covariates for the future, this is technically data leakage, and could result in an over-optimistic estimate of predictive performance. In mipdeval, covariate censoring of future covariate data is switched on by default (so no data leakage), but it can be switched off if we want to match the behavior of PsN::proseval exactly.

incremental

should MAP Bayesian do incremental fits in the iterative loop. I.e. in this case it would use the first iterations MAP Bayesian estimates as input for the second iteration, and so forth. The uncertainty around the MAP estimates would be used as the new omega matrix. This approach has been called "model predictive control (MPC)" (www.page-meeting.org/?abstract=9076) and may be more predictive than "regular" MAP in some scenarios. Default is FALSE.

threads

number of threads to divide computations on. Default is 1, i.e. no parallel execution

progress

should a progress bar be shown? Default is TRUE, but when debugging the package it is useful to have it off, since progress bar handlers obscure R output.

Value

A named list of data frames.